requirements.txt
defusedxml==0.4.1
Django==1.11
django-allauth==0.30.0
oauthlib==2.0.1
PyJWT==1.4.2
python-social-auth==0.3.3
python3-openid==3.0.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
channels
asgi_redis
pyhaikunator
django-redis
raven
Twisted
psycopg2
